backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
bcrypt>=4.0.0
google-auth>=2.29.0
